refactor, change minor parts
