+++ bin/graphql-let.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../dist/run');
+require('../dist/command');
