+++ src/lib/hash.ts
