+++ src/lib/print.ts
@@ -1,7 +1,7 @@
 const PREFIX = '[ graphql-let ]';
 
 export function printInfo(message: string): void {
-  console.info([PREFIX, message].join('\t'));
+  console.info([PREFIX, message].join(' '));
 }
 
 export function printError(err: Error): void {
